\name{blocks}
\alias{blocks}
\title{Break line of text into blocks of size blocksize}
\usage{
  blocks(line, blocksize = 10L, sep = " ")
}
\arguments{
  \item{line}{text}

  \item{blocksize}{number of characters in each block}

  \item{sep}{space character between blocks}
}
\description{
  Break line of text into blocks of size blocksize
}

