\name{genbank_seqblock}
\alias{genbank_seqblock}
\title{Generate a GenBank sequence block
    1    6     11   16    21   26    31   36    41   46    51   56
    |    |     |    |     |    |     |    |     |    |     |    |
    |||||||||| |||||||||| |||||||||| |||||||||| |||||||||| ||||||||||
  1 GATCACAGGT CTATCACCCT ATTAACCACT CACGGGAGCT CTCCATGCAT TTGGTATTTT
 61 CGTCTGGGGG GTGTGCACGC GATAGCATTG CGAGACGCTG GAGCCGGAGC ACCCTATGTC
121 GCAGTATCTG TCTTTGATTC CTGCCCCATC CTATTATTTA TCGCACCTAC GTTCAATATT
181 ACAGGCGAAC ATACTTACTA AAGTGTGTTA ATTAATTAAT GCTTGTAGGA CATAATAATA
241 ACAATTGAAT GTCTGCACAG CCGCTTTCCA CACAGACATC ATAACAAAAA ATTTCCACCA
301 AACCCCCCCT CCCCCGCTTC TGGCCACAGC ACTTAAACAC ATCTCTGCCA AACCCCAAAA
361 ACAAAGAACC CTAACACCAG CCTAACCAGA TTTCAAATTT TATCTTTTGG CGGTATGCAC
421 TTTTAACAGT CACCCCCCAA CTAACACATT ACTTTCCCCT CCCACTCCCA TACTACTAAT
481 CTCATCAATA CAACCCCCGC CCATCCTACC CAGCACACAC ACCGCTGCTA ACCCCATACC
541 CCGAACCAAC CAAACCCCAA AGACACCCCC CACAGTTTAT GTAGCTTACC TCCTCAAAGC
601 AATACACTGA AAATGTTTAG ACGGGCTCAC}
\usage{
  genbank_seqblock(string, start = 1L, end = -1L,
    blocksize = 10L, width = 60L, sep = " ",
    line_break = "\n", ruler = TRUE, ...)
}
\arguments{
  \item{start,}{end substring start and end}

  \item{blocksize}{number of characters in each block}

  \item{sep}{space character between blocks}

  \item{line_break}{the line break character(s)}

  \item{ruler}{display an ascii ruler}
}
\value{
  character
}
\description{
  Generate a GenBank sequence block 1 6 11 16 21 26 31 36
  41 46 51 56 | | | | | | | | | | | | |||||||||| ||||||||||
  |||||||||| |||||||||| |||||||||| |||||||||| 1 GATCACAGGT
  CTATCACCCT ATTAACCACT CACGGGAGCT CTCCATGCAT TTGGTATTTT 61
  CGTCTGGGGG GTGTGCACGC GATAGCATTG CGAGACGCTG GAGCCGGAGC
  ACCCTATGTC 121 GCAGTATCTG TCTTTGATTC CTGCCCCATC
  CTATTATTTA TCGCACCTAC GTTCAATATT 181 ACAGGCGAAC
  ATACTTACTA AAGTGTGTTA ATTAATTAAT GCTTGTAGGA CATAATAATA
  241 ACAATTGAAT GTCTGCACAG CCGCTTTCCA CACAGACATC
  ATAACAAAAA ATTTCCACCA 301 AACCCCCCCT CCCCCGCTTC
  TGGCCACAGC ACTTAAACAC ATCTCTGCCA AACCCCAAAA 361
  ACAAAGAACC CTAACACCAG CCTAACCAGA TTTCAAATTT TATCTTTTGG
  CGGTATGCAC 421 TTTTAACAGT CACCCCCCAA CTAACACATT
  ACTTTCCCCT CCCACTCCCA TACTACTAAT 481 CTCATCAATA
  CAACCCCCGC CCATCCTACC CAGCACACAC ACCGCTGCTA ACCCCATACC
  541 CCGAACCAAC CAAACCCCAA AGACACCCCC CACAGTTTAT
  GTAGCTTACC TCCTCAAAGC 601 AATACACTGA AAATGTTTAG
  ACGGGCTCAC
}

