% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciiruler.R
\name{print.asciiruler}
\alias{print.asciiruler}
\title{Print an asciiruler}
\usage{
\method{print}{asciiruler}(x, ...)
}
\arguments{
\item{x}{the asciiruler}

\item{...}{additional arguments}
}
\description{
Print an asciiruler
}
