% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciiruler.R
\name{width}
\alias{width}
\title{Get the total width of an ascii ruler}
\usage{
width(x, ...)
}
\arguments{
\item{x}{the asciiruler}

\item{...}{additional arguments}
}
\description{
Get the total width of an ascii ruler
}
